requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.10.2
cachetools==5.3.3
Cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.1
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.1
Flask-Session==0.6.0
google-auth==2.28.2
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.17.0
